Added markdown support
