+++ package.json
@@ -12,15 +12,13 @@
     "@types/react": "^18.2.47",
     "@types/react-dom": "^18.2.18",
     "bootstrap": "^5.3.2",
-    "ga-4-react": "^0.1.281",
     "react": "^18.2.0",
     "react-bootstrap": "^2.9.2",
     "react-dom": "^18.2.0",
-    "react-ga4": "^2.1.0",
+    "react-markdown": "^9.0.1",
     "react-router-dom": "^6.21.1",
     "react-scripts": "^5.0.1",
-    "typescript": "^4.9.5",
-    "web-vitals": "^2.1.4"
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "predeploy": "npm run build",
