+++ src/Content/Personal/PersonalContent.ts
@@ -1,0 +1,9 @@
+import {PageInfo} from "../../Common/ContentEntry"
+import mtb from "./mtb/mtb.md"
+import {ProfessionalContent} from "../Professional/ProfessionalContent"
+
+export const PersonalContent: PageInfo = {
+    heading: "Hello Markdown",
+    summary: mtb,
+    content: ProfessionalContent,
+}
