--- src/ContentEntry.d.ts
@@ -1,10 +1,0 @@
-export interface ContentEntry {
-    heading: string;
-    body: string;
-    images: string[];
-    links?: any[]
-}
-
-export interface EntryCollection {
-    entries: ContentEntry[]
-}
