+++ src/Pages/Contact.tsx
@@ -1,0 +1,37 @@
+import font from "../Styles/variables";
+import Row from "react-bootstrap/Row";
+import Col from "react-bootstrap/Col";
+
+export const Contact = () => {
+  return (
+    <>
+      <div
+        className="vh-100"
+        style={{
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "center",
+          width: "100%",
+        }}
+      >
+        <Row style={{ maxWidth: "100%" }}>
+          <Col>
+            <Row style={{ justifyContent: "center" }}>
+              <h2
+                style={{
+                  color: font.color,
+                  marginBottom: "10px",
+                  textAlign: "center",
+                  alignContent: "center",
+                  maxWidth: "90%",
+                }}
+              >
+                Reach out (under construction)
+              </h2>
+            </Row>
+          </Col>
+        </Row>
+      </div>
+    </>
+  );
+};
