+++ src/Styles/index.css
@@ -1,0 +1,3 @@
+.nav-link.active {
+  text-decoration: underline;
+}
