+++ src/Styles/variables.ts
@@ -1,0 +1,5 @@
+let font = {
+    color: "000000"
+}
+
+export default font;
