+++ src/global.d.ts
@@ -1,0 +1,4 @@
+declare module "*.md" {
+    const value: string; // markdown is just a string
+	export default value;
+}
