--- src/index.css
@@ -1,3 +1,0 @@
-.nav-link.active {
-  text-decoration: underline;
-}
