+++ src/index.tsx
@@ -1,19 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import "./index.css";
+import "./Styles/index.css";
 import "bootstrap/dist/css/bootstrap.css";
-import { HashRouter } from "react-router-dom";
 import { BrowserRouter } from "react-router-dom";
-
-// ReactDOM.render(
-//   <React.StrictMode>
-//     <HashRouter>
-//       <App />
-//     </HashRouter>
-//   </React.StrictMode>,
-//   document.getElementById("root")
-// );
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
@@ -25,15 +15,3 @@ root.render(
     </BrowserRouter>
   </React.StrictMode>
 );
-
-// ReactDOM.render(
-//   <BrowserRouter>
-//     <App />
-//   </BrowserRouter>,
-//   document.getElementById("root")
-// );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();
