--- src/variables.ts
@@ -1,7 +1,0 @@
-import { FormEncType } from "@remix-run/router"
-
-let font = {
-    color: "000000"
-}
-
-export default font;
